Movable columns in detail view
This is a mod that's been in use in Puppy Linux world for years.
It makes possible to reorder columns in detail view.
It also moves 'Size' column more to the left (to be right after 'Type').

I'm not the author of this patch, I only made it compliant with current sources.
I think it all started right here:
http://www.murga-linux.com/puppy/viewtopic.php?t=57166
and Otter & Rocket are the users responsible for this mod.

What do you think about it?
(cherry picked from commit 7391416acdc303d6ef00a8bb2d20199e1e035e48)