rox-filer-2013.6-comment.patch
Add support for comment file for files and directories. Comment is
displayed in a tooltip. Comment file is a regular text file with
the pattern of .${fname}.comment, so a file called "style.css"
will have ".style.css.comment" as its comment file.

Comment file is optional and may be empty, in which case no
tooltip will be shown. Maximum size of a comment is 1024 bytes.
There is no markup processing, text in a comment file will be
shown as is.

There is no GUI to create/edit comment files, an external utility
is needed.

Ideas from Grant (smokey01).

James 2015