run-action-helper
Allows the use of a helper script when no
run action is set for a file type.

MIME type of the file is passed to the script as $1
and the file name is passed next.

If the file name or path contain spaces, it will be broken up into
multiple arguments, and if there are multiple spaces next to each
other they will be replaced with a single space.

This is not ideal, and may need improvement.